landing tests migrated, further test cases required
